feat(!custom-formats): add support for custom filename formats
Added support for custom filename formats. This is done by allowing the user to choose a suffix, datetime format and extra characters to remove.

BREAKING CHANGE: Affects old filenames that use the exact filename format as in the `data/files_fixture.json` file. An easy fix would be to add `suffix = "."` to your `config.toml` file.